Server/op5Monitor: Avoid syntax errors when scheduling downtime
The previous code was just plain wrong, so we borrow one
function (get_start_end()) from the Multisite driver instead,
as that one seems to work just fine. In order to make sure we
still pass timestamps to the API, we convert the human-readable
values back to timestamps in the _set_downtime() method.

Signed-off-by: Andreas Ericsson <ae@op5.se>