Server/op5Monitor: Avoid sending 'author' as a parameter to API
The API requires authentication (in some form or other), so it
will always take the "author" parameter from the logged in user
instead of as a user-controlled variable. This patch removes the
offending parameters from the commands that had them, allowing
the API to stop moping and start doing some work instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>