Check_MK Multisite: Added server option to enforce AuthUser
When a user which is permitted to see all hosts/services in Check_MK
Multisite is used from Nagstamon it displays the problems of all
hosts/services.

Some admin users only want to see "their" hosts. These are the hosts
they are assigned with as contact.

Since Check_MK 1.4.0i1 the new option to "show only permitted hosts"
is supported by Check_MK.

Previous Check_MK versions had a user profile option named "Visibility of
Hosts/Services (Webservice)" to configure this. But this option has
been deprecated and will be removed in future releases. If you used this
option before you are adviced to disable it and enable the new option
in Nagstamon.