Really disable InsecureRequestWarning warnings
Because of the stubbing and linking and whatnot in requests.packages
performed by requests itself and distribution maintainers, using
disable_warnings without parameters sometimes doesn't work (e.g. on
Debian Stretch). Explicitly passing in the module reference for which to
disable warnings makes it work as intended.

References: #302 and https://bugs.debian.org/861152