Increase flicker prevention delay to .5 secs
The problem happens in Linux/KDE when the window height is bigger than the
screen height. As soon as the window is shown, it gets on top of the tray bar
and an "enterEvent fires". Moments later (sometimes even .2 secs later), the
tray bar gets back to top, the window goes behind it, and a "leaveEvent" is
fired.

A .5 sec timeout seems to be a reasonable timeout also in different situations.