3.17 20250223 (#1105)
* Alertmanager: Multiple alerts per host,service pair (#1101)

For the prometheus alertmanager server module allow multiple alerts with
the same name per host,service pair. Use the alerts' fingerprints as
'internal' name and the generic display-name in the UI.

* 3.17-20250223

* fix(alertmanager): Continue loop when severity is None (#1099)

Hello,

When the alert is on severity: None, the next alerts are not displayed.
It's probably due because the loop is exited with a break, but probably should continue to the next items

* make build.py more robust with pyinstaller

* 3.17-20250223 debug

* 3.17-20250223 no trust_env

* fix Nagstamon.app in build.py

* updated CREDITS

---------

Co-authored-by: Thomas Erbesdobler <t.erbesdobler@gmx.de>
Co-authored-by: adrian-salas <adrian.salas84@gmail.com>
Co-authored-by: Henri Wahl <henri.wahl@ukdd.de>